requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

2to3
beautifulsoup4
future
fuzzywuzzy
guesslang
html2text
pipreqs
py-stackexchange
python-dotenv
pyupgrade
requests
